trezor/utils: refactor 'serialize_identity()'
